Improve mutex & fix threaded syscall bug
Changed the reentrant mutex implementation so that it disables
interrupts while it is locked and enables them again when the last
ticket is given back (and only if interrupts where enabled beforehand).
Fixed a bug in the threaded syscall, added functionality to make sure
that "taskingSchedule" is only called during interrupt handling.